fixed non-synchronized values

+++ src/main/java/org/hashsnail/server/model/range/PasswordRange.java
@@ -57,21 +57,21 @@ public class PasswordRange {
 
         char[] edgePassword = new char[mask.length];
         int[] base = new int[mask.length];
-        int decimalRepresentation = 1;
-        int symbolIndex;
+        long decimalRepresentation = 1;
+        long symbolIndex;
 
         base[mask.length - 1] = 1;
         for (int i = mask.length - 2; i >= 0; i--) {
             base[i] = mask[i + 1].getAlphabet().length * base[i + 1];
         }
-        decimalRepresentation = base[0] * mask[0].getAlphabet().length;
+        decimalRepresentation = (long) base[0] * mask[0].getAlphabet().length;
 
         decimalRepresentation *= proportion;
 
         for(int i = 0; i < mask.length; i++) {
             symbolIndex = decimalRepresentation / base[i];
             decimalRepresentation %= base[i];
-            edgePassword[i] = mask[i].getAlphabet()[symbolIndex];
+            edgePassword[i] = mask[i].getAlphabet()[(int) symbolIndex];
         }
 
         return edgePassword;
